[Fix] Move getChat into pageController
