+++ src/controllers/pageController.ts
@@ -6,6 +6,7 @@ import { generateResponse } from '../helpers';
 import { ResponseObject, Following } from '../types';
 import { SERVER_OK, SERVER_BAD_REQUEST } from '../constants';
 import requestModel from '../models/requestModel';
+import chatModel from '../models/chatModel';
 
 async function myProfile(req: Request, res: Response) {
   try {
@@ -250,4 +251,89 @@ async function userRequest(req: Request, res: Response) {
   }
 }
 
-export default { myProfile, userProfile, home, myRequest, userRequest };
+async function getMessage(req: Request, res: Response) {
+  try {
+    let decoded = (<any>req).decoded;
+    let { id } = decoded;
+    let redundantChat = await chatModel.getGroupedChat(id);
+    let chatList: Array<{
+      user_id?: number;
+      sender_id?: number;
+      receiver_id?: number;
+    }> = [];
+    let flag: boolean;
+    redundantChat.data.forEach((firstItem) => {
+      flag = true;
+      chatList.forEach((item) => {
+        if (
+          firstItem.sender_id === item.receiver_id &&
+          firstItem.receiver_id === item.sender_id
+        ) {
+          flag = false;
+        }
+      });
+      if (flag) {
+        chatList.push(firstItem);
+      }
+    });
+
+    chatList = chatList.map((item) => {
+      if (item.sender_id === id) {
+        return { user_id: item.receiver_id };
+      } else {
+        return { user_id: item.sender_id };
+      }
+    });
+
+    let chatListAll = await Promise.all(
+      chatList.map(async (item) => {
+        let userResponse = await userModel.getUserById(item.user_id);
+        let messageList: ResponseObject = await chatModel.getChatBySenderIDOrReceiverID(
+          id,
+          item.user_id,
+        );
+        return { ...userResponse.data[0], messages: messageList.data };
+      }),
+    );
+
+    chatListAll.sort((firstItem, secondItem) => {
+      return (
+        secondItem.messages[secondItem.messages.length - 1].timestamp -
+        firstItem.messages[firstItem.messages.length - 1].timestamp
+      );
+    });
+
+    let finalResponse = {
+      success: true,
+      data: [
+        {
+          my_data: [
+            await userModel.getUserById(id).then((item) => {
+              return item.data[0];
+            }),
+          ],
+          chat_list: chatListAll,
+        },
+      ],
+      message: 'Successfully getting all of the chat list',
+    };
+
+    if (finalResponse.success) {
+      res.status(SERVER_OK).json(generateResponse(finalResponse));
+    } else {
+      res.status(SERVER_BAD_REQUEST).json(generateResponse(finalResponse));
+    }
+  } catch (e) {
+    res.status(SERVER_BAD_REQUEST).json(String(e));
+    return;
+  }
+}
+
+export default {
+  myProfile,
+  userProfile,
+  home,
+  myRequest,
+  userRequest,
+  getMessage,
+};
