+++ src/routes/pageRoute.ts
@@ -10,5 +10,6 @@ pageRouter.get('/profile/:id', pageController.userProfile);
 pageRouter.get('/home', pageController.home);
 pageRouter.get('/my-request', pageController.myRequest);
 pageRouter.get('/user-request', pageController.userRequest);
+pageRouter.get('/chat', pageController.getMessage);
 
 export default pageRouter;
